--- src/main/java/online/devupgrade/sezon2/ProductCommand.java
@@ -1,3 +1,0 @@
-package online.devupgrade.sezon2;
-public class ProductCommand {
-}
