--- src/main/java/online/devupgrade/sezon2/ProductRepo.java
@@ -1,17 +1,0 @@
-package online.devupgrade.sezon2;
-
-import java.util.List;
-
-public class ProductRepo {
-    static List<Product> loadAllProductsFromOrder(Integer orderId) {
-        //TODO sql
-        return null;
-    }
-
-    public static Object persist(Object p1) {
-        //TODO sql
-        return p1;
-    }
-
-
-}
