+++ src/main/java/online/devupgrade/sezon2/dto/ProductCommand.java
@@ -1,0 +1,3 @@
+package online.devupgrade.sezon2.dto;
+public class ProductCommand {
+}
