+++ src/main/java/online/devupgrade/sezon2/entities/DefaultStatus.java
@@ -1,4 +1,6 @@
-package online.devupgrade.sezon2;
+package online.devupgrade.sezon2.entities;
+
+import online.devupgrade.sezon2.helper.IOrderStatus;
 
 public enum DefaultStatus implements IOrderStatus {
     W_Przygotowaniu(0), Zamowione(1), Anulowane(2), Anulowane_anulowanie(3);
