+++ src/main/java/online/devupgrade/sezon2/entities/NotReallyOrderOrder.java
@@ -1,4 +1,4 @@
-package online.devupgrade.sezon2;
+package online.devupgrade.sezon2.entities;
 
 public class NotReallyOrderOrder extends Order {
     public NotReallyOrderOrder(Order order) {
