+++ src/main/java/online/devupgrade/sezon2/entities/StringBasedOrderStatus.java
@@ -1,8 +1,10 @@
-package online.devupgrade.sezon2;
+package online.devupgrade.sezon2.entities;
+
+import online.devupgrade.sezon2.helper.IOrderStatus;
 
 import java.util.stream.IntStream;
 
-public class StringBasedOrderStatus implements  IOrderStatus, CharSequence{
+public class StringBasedOrderStatus implements IOrderStatus, CharSequence{
 
     private String implHelper; //kompozycja
