+++ src/main/java/online/devupgrade/sezon2/services/FindOrdersCommandHandler.java
@@ -1,9 +1,8 @@
 package online.devupgrade.sezon2.services;
 
 import online.devupgrade.sezon2.dto.FindOrdersCommand;
-import online.devupgrade.sezon2.OrderDTO;
-import online.devupgrade.sezon2.ProductManagerLogicService;
-import online.devupgrade.sezon2.ResultTransporterException;
+import online.devupgrade.sezon2.dto.OrderDTO;
+import online.devupgrade.sezon2.dto.ResultTransporterException;
 import online.devupgrade.sezon2.utils.Command;
 import online.devupgrade.sezon2.utils.CommandHandler;
 import org.springframework.beans.factory.annotation.Autowired;
