+++ src/main/java/online/devupgrade/sezon2/utilshelpers/OrderPriceSumCalculator.java
@@ -1,4 +1,7 @@
-package online.devupgrade.sezon2;
+package online.devupgrade.sezon2.utilshelpers;
+
+import online.devupgrade.sezon2.helper.IVisitor;
+import online.devupgrade.sezon2.entities.Product;
 
 import java.util.List;
 import java.util.Map;
@@ -6,7 +9,7 @@ import java.util.Optional;
 
 public class OrderPriceSumCalculator implements IVisitor {
 
-    final SumProvider sumMantisa = new SumProvider();
+    final public SumProvider sumMantisa = new SumProvider();
 
     @Override
     public Optional<Boolean> visit(List<Product> products, Optional<Map<String, Object>> params) {
@@ -19,12 +22,4 @@ public class OrderPriceSumCalculator implements IVisitor {
         Float sum = sumMantisa.sum;
         return sum > 0f ? Optional.of(Boolean.TRUE) : Optional.empty();
     }
-}
-class SumProvider {
-    static Float sum = 0f;
-
-    public void add(Float a) {
-        sum = sum + a;
-    }
-
 }
